Rename plista-outbrain-renderer.js into plista-renderer.js
